deploy-gae/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dataclasses-json==0.5.14
dnspython==2.4.1
Flask==2.3.2
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-PyMongo==2.3.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.252
langsmith==0.0.19
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pymongo==4.4.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.7.1
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2